Move member functions back to secretcache.go + fix lint
