+++ security/pkg/nodeagent/cache/secretcache.go
@@ -20,10 +20,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"math/rand"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
+
+	"github.com/gogo/status"
 
 	"istio.io/istio/security/pkg/nodeagent/model"
 	"istio.io/istio/security/pkg/nodeagent/plugin"
@@ -245,6 +248,63 @@ func (sc *SecretCache) GenerateSecret(ctx context.Context, connectionID, resourc
 	return ns, nil
 }
 
+// SecretExist checks if secret already existed.
+// This API is used for sds server to check if coming request is ack request.
+func (sc *SecretCache) SecretExist(connectionID, resourceName, token, version string) bool {
+	key := ConnKey{
+		ConnectionID: connectionID,
+		ResourceName: resourceName,
+	}
+	val, exist := sc.secrets.Load(key)
+	if !exist {
+		return false
+	}
+
+	e := val.(model.SecretItem)
+	return e.ResourceName == resourceName && e.Token == token && e.Version == version
+}
+
+// IsIngressGatewaySecretReady returns true if node agent is working in ingress gateway agent mode
+// and needs to wait for ingress gateway secret to be ready.
+func (sc *SecretCache) ShouldWaitForIngressGatewaySecret(connectionID, resourceName, token string) bool {
+	// If node agent works as workload agent, node agent does not expect any ingress gateway secret.
+	if sc.fetcher.UseCaClient {
+		return false
+	}
+
+	key := ConnKey{
+		ConnectionID: connectionID,
+		ResourceName: resourceName,
+	}
+	// Add an entry into cache, so that when ingress gateway secret is ready, gateway agent is able to
+	// notify the ingress gateway and push the secret to via connect ID.
+	if _, found := sc.secrets.Load(key); !found {
+		t := time.Now()
+		dummySecret := &model.SecretItem{
+			ResourceName: resourceName,
+			Token:        token,
+			CreatedTime:  t,
+			Version:      t.String(),
+		}
+		sc.secrets.Store(key, *dummySecret)
+	}
+
+	conIDresourceNamePrefix := cacheLogPrefix(connectionID, resourceName)
+	// If node agent works as ingress gateway agent, searches for kubernetes secret and verify secret
+	// is not empty.
+	cacheLog.Debugf("%s calling SecretFetcher to search for secret %s",
+		conIDresourceNamePrefix, resourceName)
+	_, exist := sc.fetcher.FindIngressGatewaySecret(resourceName)
+	// If kubernetes secret does not exist, need to wait for secret.
+	if !exist {
+		cacheLog.Warnf("%s SecretFetcher cannot find secret %s from cache",
+			conIDresourceNamePrefix, resourceName)
+		return true
+	}
+
+	return false
+}
+
 // DeleteSecret deletes a secret by its key from cache.
 func (sc *SecretCache) DeleteSecret(connectionID, resourceName string) {
 	key := ConnKey{
@@ -254,9 +314,46 @@ func (sc *SecretCache) DeleteSecret(connectionID, resourceName string) {
 	sc.secrets.Delete(key)
 }
 
+func (sc *SecretCache) callbackWithTimeout(connectionID string, secretName string, secret *model.SecretItem) {
+	c := make(chan struct{})
+	conIDresourceNamePrefix := cacheLogPrefix(connectionID, secretName)
+	go func() {
+		defer close(c)
+		if sc.notifyCallback != nil {
+			if err := sc.notifyCallback(connectionID, secretName, secret); err != nil {
+				cacheLog.Errorf("%s failed to notify secret change for proxy: %v",
+					conIDresourceNamePrefix, err)
+			}
+		} else {
+			cacheLog.Warnf("%s secret cache notify callback isn't set", conIDresourceNamePrefix)
+		}
+	}()
+	select {
+	case <-c:
+		return // completed normally
+	case <-time.After(notifyK8sSecretTimeout):
+		cacheLog.Warnf("%s notify secret change for proxy got timeout", conIDresourceNamePrefix)
+	}
+}
+
 // Close shuts down the secret cache.
 func (sc *SecretCache) Close() {
 	sc.closing <- true
+}
+
+func (sc *SecretCache) keyCertRotationJob() {
+	// Wake up once in a while and refresh stale items.
+	sc.rotationTicker = time.NewTicker(sc.configOptions.RotationInterval)
+	for {
+		select {
+		case <-sc.rotationTicker.C:
+			sc.rotate(false /*updateRootFlag*/)
+		case <-sc.closing:
+			if sc.rotationTicker != nil {
+				sc.rotationTicker.Stop()
+			}
+		}
+	}
 }
 
 // DeleteK8sSecret deletes all entries that match secretName. This is called when a K8s secret
@@ -564,4 +661,98 @@ func (sc *SecretCache) generateSecret(ctx context.Context, token, resourceName s
 		ExpireTime:       expireTime,
 		Version:          t.String(),
 	}, nil
+}
+
+func (sc *SecretCache) shouldRefresh(s *model.SecretItem) bool {
+	// secret should be refreshed before it expired, SecretRefreshGraceDuration is the grace period;
+	return time.Now().After(s.ExpireTime.Add(-sc.configOptions.SecretRefreshGraceDuration))
+}
+
+func (sc *SecretCache) isTokenExpired() bool {
+	// skip check if the token passed from envoy is always valid (ex, normal k8s sa JWT).
+	if sc.configOptions.AlwaysValidTokenFlag {
+		return false
+	}
+
+	if atomic.LoadUint32(&sc.skipTokenExpireCheck) == 1 {
+		return true
+	}
+	// TODO(quanlin), check if token has expired.
+	return false
+}
+
+// sendRetriableRequest sends retriable requests for either CSR or ExchangeToken.
+// Prior to sending the request, it also sleep random millisecond to avoid thundering herd problem.
+func (sc *SecretCache) sendRetriableRequest(ctx context.Context, csrPEM []byte, providedExchangedToken, resourceName string, isCSR bool) ([]string, error) {
+	backOffInMilliSec := rand.Int63n(sc.configOptions.InitialBackoff)
+	cacheLog.Debugf("Wait for %d millisec for initial CSR", backOffInMilliSec)
+	// Add a jitter to initial CSR to avoid thundering herd problem.
+	time.Sleep(time.Duration(backOffInMilliSec) * time.Millisecond)
+
+	startTime := time.Now()
+	var retry int64
+	var certChainPEM []string
+	exchangedToken := providedExchangedToken
+	var requestErrorString string
+	var err error
+
+	// Keep trying until no error or timeout.
+	for {
+		var httpRespCode int
+		if isCSR {
+			requestErrorString = fmt.Sprintf("CSR for %q", resourceName)
+			certChainPEM, err = sc.fetcher.CaClient.CSRSign(
+				ctx, csrPEM, exchangedToken, int64(sc.configOptions.SecretTTL.Seconds()))
+		} else {
+			requestErrorString = "Token exchange"
+			p := sc.configOptions.Plugins[0]
+			exchangedToken, _, httpRespCode, err = p.ExchangeToken(ctx, sc.configOptions.TrustDomain, exchangedToken)
+		}
+
+		if err == nil {
+			break
+		}
+
+		// If non-retryable error, fail the request by returning err
+		if !isRetryableErr(status.Code(err), httpRespCode, isCSR) {
+			cacheLog.Errorf("%s hit non-retryable error %v", requestErrorString, err)
+			return nil, err
+		}
+
+		// If reach envoy timeout, fail the request by returning err
+		if startTime.Add(time.Millisecond * envoyDefaultTimeoutInMilliSec).Before(time.Now()) {
+			cacheLog.Errorf("%s retry timed out %v", requestErrorString, err)
+			return nil, err
+		}
+
+		retry++
+		backOffInMilliSec = rand.Int63n(retry * initialBackOffIntervalInMilliSec)
+		time.Sleep(time.Duration(backOffInMilliSec) * time.Millisecond)
+		cacheLog.Warnf("%s failed with error: %v, retry in %d millisec", requestErrorString, err, backOffInMilliSec)
+	}
+
+	if isCSR {
+		return certChainPEM, nil
+	}
+	return []string{exchangedToken}, nil
+}
+
+// getExchangedToken gets the exchanged token for the CSR. The token is either the k8s jwt token of the
+// workload or another token from a plug in provider.
+func (sc *SecretCache) getExchangedToken(ctx context.Context, k8sJwtToken string) (string, error) {
+	exchangedTokens := []string{k8sJwtToken}
+	var err error
+	if sc.configOptions.Plugins != nil && len(sc.configOptions.Plugins) > 0 {
+		// Currently the only plugin is GoogleTokenExchange, so we only extract the first possible plugin.
+		if len(sc.configOptions.Plugins) > 1 {
+			cacheLog.Error("found more than one plugin")
+			return "", err
+		}
+		exchangedTokens, err = sc.sendRetriableRequest(ctx, nil, k8sJwtToken, "", false)
+		if err != nil || len(exchangedTokens) == 0 {
+			cacheLog.Errorf("failed to exchange token: %v", err)
+			return "", err
+		}
+	}
+	return exchangedTokens[0], nil
 }
